Use table more directly
